restructuring with unittest complete!
