--- model/conf/config.yml
@@ -1,25 +1,0 @@
-unittest_settings:
-  input_dimension: [1,4]    # specify dimension of input
-  input_type: "number"    # choose from ["number" | "text"]
-model_settings:
-  gpu: false
-  pretrained_filepath: "./saved_models/torch_example.pt"
-  # pretrained_filepath: "/home/eppen/kpmg/corona-blue/Emotion-detection/mymodel"
-
-aws_settings:
-  sqs:
-    request:
-      audio:
-      nlp:
-      video:
-      vision:
-    result:
-      audio:
-      nlp:
-      video:
-      vision:
-  ecr:
-    audio:
-    nlp:
-    video:
-    vision:
