--- model/conf/requirements.txt
@@ -1,4 +1,0 @@
-tensorflow==2.2.0
-numpy==1.17.4
-torch==1.7.1
-PyYAML==5.4.1
