--- model/src/fer_model/ml/trained_models/esr_9/Net-Base-Shared_Representations.pt
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:563215f7ae9487277f2da2f1b48d6d1e3b3de5daa1cf377429dd5368379acb97
-size 1506551
