--- model/src/fer_model/ml/trained_models/esr_9/Net-Branch_1.pt
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:93f318a411f6447d426380c11fa3e6edad64ce277ab99a895535a3c8a4394302
-size 8891474
