--- model/src/fer_model/ml/trained_models/esr_9/Net-Branch_2.pt
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:c16882cfe35fce26079cf98f397f3f3a231975d4fb621de7e573e101ffbfd664
-size 8891466
