--- model/src/fer_model/ml/trained_models/esr_9/Net-Branch_3.pt
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:2156a70028481606bd6278c58c5270766a2b7cc17ff9f00d36bf8134a2f20eff
-size 8891462
