--- model/src/fer_model/ml/trained_models/esr_9/Net-Branch_4.pt
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:7570df10b23348b0e22ab41269e93883fb8c9449fb4e7200cc1de27709fd5569
-size 8891458
