--- model/src/fer_model/ml/trained_models/esr_9/Net-Branch_5.pt
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:04483ba790d9696a2254e8d4f3dbfdd5407fb2516ee0e8fff52f108b17becbca
-size 8891474
