--- model/src/fer_model/ml/trained_models/esr_9/Net-Branch_6.pt
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:71eca653bb8976b1eea0e8660f238dabcdc68a1ccb6ab6e19e6c835c1b0559b4
-size 8891466
