--- model/src/fer_model/ml/trained_models/esr_9/Net-Branch_7.pt
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:d28ce6639c73cc9c93ef20777a74c35a35c8ad10bb5738c891b93472ede05d7f
-size 8891474
