--- model/src/fer_model/ml/trained_models/esr_9/Net-Branch_8.pt
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:048dbf5da4730b39f35a58cdc4caabea104e220b9e6ef17a04c9c949f9cea5fc
-size 8891470
