--- model/src/fer_model/ml/trained_models/esr_9/Net-Branch_9.pt
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:14df8cc559ab4f63dd8358523d33c5457f698e1563344fbb0dd8646f4b7fac68
-size 8891478
