+++ video_model/conf/requirements.txt
@@ -1,0 +1,4 @@
+tensorflow==2.2.0
+numpy==1.17.4
+torch==1.7.1
+PyYAML==5.4.1
